+++ samples/ZenUI.Wpf.Gallery/Views/NumberBoxView.xaml.cs
@@ -1,4 +1,6 @@
 using System.Windows.Controls;
+using System.Windows.Input;
+using Prism.Commands;
 
 namespace ZenUI.Wpf.Gallery.Views
 {
@@ -6,7 +8,15 @@ namespace ZenUI.Wpf.Gallery.Views
     {
         public NumberBoxView()
         {
+            EditorFeedbackCommand = new DelegateCommand<object>(ShowEditorCommandFeedback);
             InitializeComponent();
+        }
+
+        public ICommand EditorFeedbackCommand { get; }
+
+        private void ShowEditorCommandFeedback(object parameter)
+        {
+            EditorCommandStatus.Text = "已执行命令，参数：" + (parameter ?? "null");
         }
     }
 }