+++ src/ZenUI.Wpf/Controls/RadioGroupAppearance.cs
@@ -1,9 +1,9 @@
 namespace ZenUI.Wpf.Controls
 {
     /// <summary>
-    /// 指定 <see cref="ZenRadioGroup"/> 选项的显示模式。
+    /// 指定 <see cref="ZenRadioGroup"/> 选项的视觉外观。
     /// </summary>
-    public enum RadioGroupDisplayMode
+    public enum RadioGroupAppearance
     {
         /// <summary>
         /// 使用传统单选按钮呈现选项。
@@ -18,7 +18,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 使用描边按钮呈现选项。
         /// </summary>
-        Outline,
+        Outlined,
 
         /// <summary>
         /// 使用无常驻背景和边框的按钮呈现选项。
@@ -33,6 +33,6 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 使用相连的胶囊式分段控件呈现选项。
         /// </summary>
-        Segment
+        Segmented
     }
 }